removed sdk flow
