+++ config/docker_compose.toml
@@ -123,8 +123,7 @@ bitpay.base_url = "https://test.bitpay.com"
 bluesnap.base_url = "https://sandbox.bluesnap.com/"
 bluesnap.secondary_base_url = "https://sandpay.bluesnap.com/"
 boku.base_url = "https://$-api4-stage.boku.com"
-braintree.base_url = "https://api.sandbox.braintreegateway.com/"
-braintree.secondary_base_url = "https://payments.sandbox.braintree-api.com/graphql"
+braintree.base_url = "https://payments.sandbox.braintree-api.com/graphql"
 cashtocode.base_url = "https://cluster05.api-test.cashtocode.com"
 checkout.base_url = "https://api.sandbox.checkout.com/"
 coinbase.base_url = "https://api.commerce.coinbase.com"
@@ -439,8 +438,6 @@ connector_list = "stripe,adyen,cybersource"
 connector_list = "gocardless,stax,stripe"
 payout_connector_list = "stripe,wise"
 
-[multiple_api_version_supported_connectors]
-supported_connectors = "braintree"
 
 [payment_method_auth]
 redis_expiry = 900
